Fix bug in calculation of the closest normal direction.
We compare absolute values, but then store the plain value as the current max. This breaks in the case where the closest off is a negative dot product, but is overridden by a subsequent, smaller positive dot product.

How did I find this? Well it apparently happened enough times in game for me to notice that the arrow was often pointing in the wrong direction.